+++ src/test/java/org/folio/rest/api/HoldingsStorageTest.java
@@ -5,6 +5,7 @@ import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
+import lombok.SneakyThrows;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -559,28 +560,34 @@ public class HoldingsStorageTest extends TestBaseWithInventoryUtil {
     assertRemoveAllEventForHolding();
   }
 
+  @SneakyThrows
   @Test
-  public void canDeleteHoldingsByCql() throws Exception {
+  public void canDeleteHoldingsByCql() {
     UUID instanceId1 = UUID.randomUUID();
     UUID instanceId2 = UUID.randomUUID();
     instancesClient.create(smallAngryPlanet(instanceId1));
     instancesClient.create(nod(instanceId2));
 
     var h1 = holdingsClient.create(new HoldingRequestBuilder()
-        .forInstance(instanceId1).withPermanentLocation(mainLibraryLocationId).withHrid("1234")
-        ).getJson();
+        .forInstance(instanceId1)
+        .withPermanentLocation(mainLibraryLocationId)
+        .withHrid("1234")).getJson();
     var h2 = holdingsClient.create(new HoldingRequestBuilder()
-        .forInstance(instanceId1).withPermanentLocation(mainLibraryLocationId).withHrid("21234")
-        ).getJson();
+        .forInstance(instanceId1)
+        .withPermanentLocation(mainLibraryLocationId)
+        .withHrid("21234")).getJson();
     var h3 = holdingsClient.create(new HoldingRequestBuilder()
-        .forInstance(instanceId2).withPermanentLocation(mainLibraryLocationId).withHrid("12")
-        ).getJson();
+        .forInstance(instanceId2)
+        .withPermanentLocation(mainLibraryLocationId)
+        .withHrid("12")).getJson();
     var h4 = holdingsClient.create(new HoldingRequestBuilder()
-        .forInstance(instanceId2).withPermanentLocation(mainLibraryLocationId).withHrid("3123")
-        ).getJson();
+        .forInstance(instanceId2)
+        .withPermanentLocation(mainLibraryLocationId)
+        .withHrid("3123")).getJson();
     var h5 = holdingsClient.create(new HoldingRequestBuilder()
-        .forInstance(instanceId2).withPermanentLocation(mainLibraryLocationId).withHrid("123")
-        ).getJson();
+        .forInstance(instanceId2)
+        .withPermanentLocation(mainLibraryLocationId)
+        .withHrid("123")).getJson();
 
     var response = client.delete(holdingsStorageUrl("?query=hrid==12*"), StorageTestSuite.TENANT_ID).get(5, SECONDS);
 
@@ -595,8 +602,9 @@ public class HoldingsStorageTest extends TestBaseWithInventoryUtil {
     assertRemoveEventForHolding(h5);
   }
 
+  @SneakyThrows
   @Test
-  public void cannotDeleteHoldingsWithEmptyCql() throws Exception {
+  public void cannotDeleteHoldingsWithEmptyCql() {
 
     var response = client.delete(holdingsStorageUrl("?query="), StorageTestSuite.TENANT_ID).get(5, SECONDS);