+++ src/test/java/org/folio/services/domainevent/DomainEventRawTest.java
@@ -8,7 +8,7 @@ import org.junit.Test;
 public class DomainEventRawTest {
 
   @Test
-  public void testUpdateEvent() {
+  public void updateEventIncludesOldAndNewRecordRespresentationsAndEventTypeAndTenant() {
     assertThat(DomainEventRaw.updateEvent("myold", "mynew", "mytenant").toString(),
         endsWith("[oldEntity=myold,newEntity=mynew,type=UPDATE,tenant=mytenant]"));
   }